changed task resource quantity tye to double

+++ health-services/transformer/src/main/java/org/egov/transformer/models/downstream/ProjectTaskIndexV1.java
@@ -38,7 +38,7 @@ public class ProjectTaskIndexV1 {
     @JsonProperty("productVariant")
     private String productVariant;
     @JsonProperty("quantity")
-    private Long quantity;
+    private Double quantity;
     @JsonProperty("deliveredTo")
     private String deliveredTo;
     @JsonProperty("isDelivered")
